Move code out of __init__.py
This makes it MUCH easier to troubleshoot with djangocms-helper
