+++ test_requirements.txt
@@ -11,7 +11,7 @@ django-reversion>=1.8.2,<1.9
 django-sekizai
 django-sortedm2m
 djangocms-admin-style
--e git+https://github.com/nephila/djangocms-helper.git#egg=djangocms-helper
+djangocms-helper
 djangocms_text_ckeditor
 docopt
 html5lib
